uint256<-->uint128<-->int128

+++ contracts/test/VeBalMock.sol
@@ -18,10 +18,10 @@ interface SmartWalletChecker {
 }   
 
 contract VeBalMock is ERC20, ReentrancyGuard {
-
+    
     struct Point{
-        int256 bias;
-        int256 slope; // - dweight / dt
+        int128 bias;
+        int128 slope; // - dweight / dt
         uint256 ts;
         uint256 blk; // block
     }
@@ -30,7 +30,7 @@ contract VeBalMock is ERC20, ReentrancyGuard {
     // What we can do is to extrapolate ***At functions
 
     struct LockedBalance{
-        int256 amount; 
+        int128 amount; 
         uint256 end;
     }  
 
@@ -68,7 +68,7 @@ contract VeBalMock is ERC20, ReentrancyGuard {
     Point[100000000000000000000000000000] public point_history; //epoch -> unsigned point
     mapping(address => Point[1000000000]) private user_point_history; //user -> Point[user_epoch]
     mapping(address => uint256) public user_point_epoch;
-    mapping(uint256 => int256) public slope_changes; //time -> signed slope change
+    mapping(uint256 => int128) public slope_changes; //time -> signed slope change
 
     // Checker for whitelisted (smart contract) wallets which are allowed to deposit
     // The goal is to prevent tokenizing the escrow
@@ -139,7 +139,7 @@ contract VeBalMock is ERC20, ReentrancyGuard {
             // raise "Smart contract depositors not allowed";
         }
     }    
-    function get_last_user_slope(address addr) external view returns (int256){
+    function get_last_user_slope(address addr) external view returns (int128){
         uint256 uepoch = user_point_epoch[addr];
         return user_point_history[addr][uepoch].slope;
     }
@@ -155,19 +155,19 @@ contract VeBalMock is ERC20, ReentrancyGuard {
         Point memory u_old; //empty(Point);
         Point memory u_new; //empty(Point);
 
-        int256 old_dslope = 0;
-        int256 new_dslope = 0;
+        int128 old_dslope = 0;
+        int128 new_dslope = 0;
         uint256 _epoch = epoch;
 
         if (addr != ZERO_ADDRESS) {
             // Calculate slopes and biases
             // Kept at zero when they have to
             if (old_locked.end > block.timestamp && old_locked.amount > 0) {
-                u_old.slope = old_locked.amount/ (int256(MAXTIME));
-                u_old.bias = u_old.slope * (int256(old_locked.end - block.timestamp));
+                u_old.slope = old_locked.amount / (int128(uint128(MAXTIME)));
+                u_old.bias = u_old.slope * (int128(uint128(old_locked.end) - uint128(block.timestamp)));
             if (new_locked.end > block.timestamp && new_locked.amount > 0) {
-                u_new.slope = new_locked.amount / (int256(MAXTIME));
-                u_new.bias = u_new.slope * (int256(new_locked.end - block.timestamp));
+                u_new.slope = new_locked.amount / (int128(uint128(MAXTIME)));
+                u_new.bias = u_new.slope * (int128(uint128(new_locked.end) - uint128(block.timestamp)));
             }
             // Read values of scheduled changes in the slope
             // old_locked.end can be in the past and in the future
@@ -203,13 +203,13 @@ contract VeBalMock is ERC20, ReentrancyGuard {
             // Hopefully it won't happen that this won't get used in 5 years!
             // If it does, users will be able to withdraw but vote weight will be broken
             t_i += WEEK;
-            int256 d_slope = 0;
+            int128 d_slope = 0;
             if (t_i > block.timestamp) {
                 t_i = block.timestamp;
             } else {
                 d_slope = slope_changes[t_i];
             }
-            last_point.bias -= last_point.slope * (int256(t_i - last_checkpoint));
+            last_point.bias -= last_point.slope * (int128(uint128(t_i - last_checkpoint)));
             last_point.slope += d_slope;
             if (last_point.bias < 0) {  // This can happen
                 last_point.bias = 0;
@@ -283,7 +283,7 @@ contract VeBalMock is ERC20, ReentrancyGuard {
         supply = supply_before + _value;
         LockedBalance memory old_locked = LockedBalance({amount : _locked.amount, end : _locked.end});//_locked;
         // Adding to existing lock, or if a lock is expired - creating a new one
-        _locked.amount += int256(_value);
+        _locked.amount += int128(uint128(_value));
         if (unlock_time != 0) {
             _locked.end = unlock_time;
         }
@@ -355,7 +355,7 @@ contract VeBalMock is ERC20, ReentrancyGuard {
         LockedBalance memory _locked = locked[msg.sender];
 
         require(block.timestamp >= _locked.end, "The lock didn't expire");
-        uint256 value = uint256(_locked.amount);
+        uint256 value = uint256(uint128(_locked.amount));
 
         LockedBalance memory old_locked = _locked;
         _locked.end = 0;
@@ -494,12 +494,12 @@ contract VeBalMock is ERC20, ReentrancyGuard {
             return 0;
         } else {
             Point memory last_point = user_point_history[addr][_epoch];
-            last_point.bias -= last_point.slope * (int256(_t - last_point.ts));
+            last_point.bias -= last_point.slope * (int128(uint128(_t - last_point.ts)));
             if (last_point.bias < 0) {
                 last_point.bias = 0;
             }
 
-            return uint256(last_point.bias);
+            return uint256(uint128(last_point.bias));
         }
     }
 
@@ -526,12 +526,12 @@ contract VeBalMock is ERC20, ReentrancyGuard {
             return 0;
         } else {
             Point memory last_point = user_point_history[addr][_epoch];
-            last_point.bias -= last_point.slope * (int256(_t - last_point.ts));
+            last_point.bias -= last_point.slope * (int128(uint128(_t - last_point.ts)));
             if (last_point.bias < 0) {
                 last_point.bias = 0;
             }
 
-            return uint256(last_point.bias);
+            return uint256(uint128(last_point.bias));
         }
     }
 
@@ -558,9 +558,9 @@ contract VeBalMock is ERC20, ReentrancyGuard {
         if (d_block != 0) {
             block_time += d_t * (_block - point_0.blk) / d_block;
         }
-        upoint.bias -= upoint.slope * (int256(block_time - upoint.ts));
+        upoint.bias -= upoint.slope * (int128(uint128(block_time - upoint.ts)));
         if (upoint.bias >= 0) {
-            return uint256(upoint.bias);
+            return uint256(uint128(upoint.bias));
         } else {
             return 0;
         }
@@ -577,13 +577,13 @@ contract VeBalMock is ERC20, ReentrancyGuard {
         uint256 t_i = (last_point.ts / WEEK) * WEEK;
         for (uint i; i < 255; i++) {
             t_i += WEEK;
-            int256 d_slope = 0;
+            int128 d_slope = 0;
             if (t_i > t) {
                 t_i = t;
             } else {
                 d_slope = slope_changes[t_i];
             }
-            last_point.bias -= last_point.slope * (int256(t_i - last_point.ts));
+            last_point.bias -= last_point.slope * (int128(uint128(t_i - last_point.ts)));
             if (t_i == t) {
                 break;
             }
@@ -593,7 +593,7 @@ contract VeBalMock is ERC20, ReentrancyGuard {
         if (last_point.bias < 0) {
             last_point.bias = 0;
         }
-        return uint256(last_point.bias);    
+        return uint256(uint128(last_point.bias));
     }
 
     /**
